fix: resolve all 86 SonarCloud issues
- S8417: remove explicit memory_order_relaxed (use default seq_cst)
  across FlowAnalysisWorker, LiveDetectionPipeline, ConsoleAlertSink,
  SignalHandler
- S7172: use has_value() for std::expected<void,E> checks
- S6045: use transparent std::equal_to<> on string containers
- S3230: move member init from ctor list to in-class initializers
- S5817: mark const-eligible methods as const
- S1048: wrap destructor bodies in try/catch
- S1905: remove redundant static_casts
- S1066: merge nested if-statements
- S1238: pass large objects by const ref in lambdas
- S6004: use if-init statements for local variables
- S859: suppress necessary const_cast with NOSONAR
- S878: replace comma operator with explicit statements
- S886: suppress intentional argv-parsing loop
- S954: move includes above forward declarations
- S1186: document intentionally empty stop() override
- S3358: replace nested ternary with switch
- S6165: replace manual erase loop with std::erase_if
- S6010: use std::filesystem::path for loadFeeds parameter
- S134: flatten deep nesting with early returns
- Accept via MCP: S5008 (PcapPlusPlus void* callback),
  S5421 (signal handler atomic), S6018 (Meyers singleton)

+++ src/ui/FlowTableModel.h
@@ -42,8 +42,7 @@ public:
   };
 
   /// Total number of columns (integral constant for use in array sizes).
-  static constexpr int kColumnCount =
-      static_cast<int>(std::to_underlying(Column::ColumnCount));
+  static constexpr int kColumnCount = std::to_underlying(Column::ColumnCount);
 
   /// Per-flow data: detection result + connection metadata.
   struct FlowRow {
